[SERVICES] Add non-integer frame rate support
Add support for non-integer frame rates like 59.94 fps (60000/1001),
29.97 fps (30000/1001), and 23.976 fps (24000/1001) for ST 2110-20,
ST 2110-30, and ST 2110-40 streams.

Changes:
- Add non-integer frame rates to SUPPORTED_VIDEO_FRAME_RATES
- Change ticks_per_media_unit from double to Rational for exact
  fractional arithmetic without cumulative drift
- Change RTPPacketContext::timestamp from uint32_t to Rational for
  precise accumulation across frames
- Update all packet buffer writers (2110-20, 2110-30, 2110-40) to use
  Rational timestamp arithmetic
- Update closed caption encoder to use exact FrameRate matching for
  CDP frame rate codes per SMPTE ST 334-2
- Update CLI help text to indicate fractional format support

This fixes RTP timestamp drift that occurred with non-integer frame
rates when truncating fractional ticks_per_media_unit values every
frame.

+++ source/services/ulp_packet_buffer_wr/writers/rtp_smpte_2110_20_packet_buffer_writer.cpp
@@ -1,6 +1,6 @@
 /*
  * SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
- * Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -59,8 +59,8 @@ inline void RTP_SMPTE_2110_20_PacketBufferWriter::update_in_media_unit_state(siz
     }
 
     if (m_rtp_packet_context->counter == video_settings.packets_in_media_unit) {
-        // ST2210-20: the timestamp SHOULD be the same for each packet of the frame/field.
-        m_rtp_packet_context->timestamp += static_cast<uint32_t>(video_settings.ticks_per_media_unit);
+        // ST2110-20: the timestamp SHOULD be the same for each packet of the frame/field.
+        m_rtp_packet_context->timestamp += video_settings.ticks_per_media_unit;
         m_rtp_packet_context->counter = 0;
         if (video_settings.video_scan_type == VideoScanType::Interlaced) {
             m_rtp_packet_context->rtp_interlace_field_indicator = !m_rtp_packet_context->rtp_interlace_field_indicator;
